+++ splunk_eventgen/lib/plugins/output/metric_httpevent.py
@@ -1,13 +1,13 @@
-from .httpevent_core import HTTPCoreOutputPlugin
+from httpevent_core import HTTPCoreOutputPlugin
 from logging_config import logger
 
-try:
-    import requests
-    from requests import Session
-    from requests_futures.sessions import FuturesSession
-    from concurrent.futures import ThreadPoolExecutor
-except ImportError:
-    pass
+# try:
+#     import requests
+#     from requests import Session
+#     from requests_futures.sessions import FuturesSession
+#     from concurrent.futures import ThreadPoolExecutor
+# except:
+#     pass
 try:
     import ujson as json
 except ImportError:
